Separate flags for is_terminal and supports_color.
On Windows, is_terminal indicates a console, which needs to be handled
specially by writeUTF8().  supports_color might be false if we are
unable to set ENABLE_VIRTUAL_TERMINAL_PROCESSING on the console.
(Such as if we are on a version of Windows before Windows 10.)

On UNIX, supports_color is always the same is is_terminal, because
there isn't a good way to know if the terminal supports color, without
going full blown termcap/terminfo, which seems like overkill.  Assume
all modern terminals support ANSI escapes for color.